Выносит функцию составления списка жанров в селектор
