+++ src/reducer/data/selector.js
@@ -2,6 +2,8 @@ import {NameSpace} from "./../name-space.js";
 import {Genres} from "./../../utils.js";
 import {createSelector} from "reselect";
 
+const MAX_NUMBER_GENRES = 10;
+
 const getFilms = (state) => {
   return state[NameSpace.DATA].films;
 };
@@ -57,6 +59,19 @@ const getFlagReviewsFetching = (state) => {
   return state[NameSpace.DATA].isReviewsFetching;
 };
 
+const getGenresList = createSelector(
+    getFilms,
+    (films) => {
+      const genres = new Set();
+      genres.add(`All genres`);
+      films.forEach((film) => {
+        genres.add(film.genre);
+      });
+      const genresList = Array.from(genres).slice(0, MAX_NUMBER_GENRES);
+      return genresList;
+    }
+);
+
 
 export {
   getFilms,
@@ -69,5 +84,6 @@ export {
   getFilmById,
   getFavoriteFilms,
   getReviews,
-  getFlagReviewsFetching
+  getFlagReviewsFetching,
+  getGenresList
 };
